Option to easilt run w/o web-host or heroku
If you want to demo this in an isolated environment or you want to run this on a disconnected network (such as for an exercise or drill in your tabletop / premortem command center), you can build a docker container that includes a minimal HTTP server and the full pewpew setup before you go offline, and run the docker container to present the solution. This is a Q&D solution for those who don't want to mess with getting on a web-host or messing with Heroku or other code-host services. 


This includes a Dockerfile and a README file to explain how to use it